Force-close insecure agent and isolate lifecycle tests
Address review feedback on the insecure HTTP wait Agent:

- Use Agent.destroy() instead of close() on teardown. Status-only
  predicate paths never consume the response body, so a graceful close()
  could hang waiting for those connections to drain. The wait has already
  finished at that point, so a forceful destroy is correct.
- Move the agent-lifecycle tests into their own file with hoisted,
  top-level mocks for undici and the container runtime, importing
  HttpWaitStrategy directly. Previously the vi.doMock("../container-runtime")
  ran after the static "./wait" import had already cached HttpWaitStrategy,
  so the mock was a no-op and the "Docker-free" tests still required a real
  container runtime. The dedicated file makes them genuinely Docker-free.
